feat(wip): tentatively complete home page
Tentatively completes the text and media content for the Home page,
which encompasses a short about section and a highlighted list of
services offered.

TODO: replace placeholder image in About page with an image of Randall,
and add links in service sections to Services sections after creating
the Services page.