fix(wip): ensure image paths work in prod
The problem: I was constructing image paths dynamically in ImagePanel
and SimpleImagePanel by passing in only path relative to
/src/assets/img, and using something like `'/src/assets/img/' + imgSrc`.
This worked locally but failed in production because Vite didn't pick up
on the dynamically required assets.

I now use full paths for all images in an attempt. This still doesn't
help with Vite picking up on the assets, but the standardization is a
step in the right direction.