fix: fix metadata-rotated images
Context: some images were previously rotated in metadata using EXIF
tags (to correct for rotated camera when the image was taken), which
then causes problems on further manipulation with imagetools, which
seems to ignore the EXIF metadata and shows the original image without
rotation.

This commit replaces all of the metadata-rotated images with equivalents
in which the actual pixels are rotated with imagemagick.